Rename function relay_update
Rename function to update_command_output() to better reflect that this
callback updates the UI with the latest output read from the command
being executed.

This makes more sense knowing a future change will add a separate
callback which parses the output read from the command and updates the
progress bar.  This function should be named update_command_progress().