Add testbuild.sh script (#699881)
For developers to build GParted in a git repository, logging the top
commit and build results to testbuild.log.  Intended for use with
git-test-sequence to verify every change of a patch set compiles, but
can be used standalone too.

Example usage:
    git-test-sequence origin/master.. testbuild.sh

Further documentation can be found on the GParted web site at page:
    Git Source Code Management
    http://www.gparted.org/git.php

Closes Bug #699881 - testbuild.sh - Builds GParted logging results