Allow GParted to read the UUID from older btrfs-show command
All the code was there in btrfs::read_uuid() to read the UUID from the
btrfs-show command.  It just needed enabling.

This is only relevant when: (1) the blkid command is unavailable as
GParted primarily reads the UUID through the FS_Info cache using the
blkid command, and (2) the btrfs command is unavailable and the btrfs
module uses the older btrfs-show command instead.