Make resize of whole disk file systems work (#743181)
Only allow resizing, not moving of a whole disk device file system.
There is no actual partition to move and moving a file system away from
the start of a disk only makes it unrecognisable.

Also don't perform the partition resize step as there's no actual
partition to be resized.  Only the file system is being resized.
(Libparted actually allows the virtual partition spanning a whole disk
device to be resized, implementing it as a no-operation, but only for
recognised file systems.  For unrecognised file systems it fails with
"unrecognised disk label").

Note that the existing resize dialog was designed for resizing partition
boundaries, and their contained file systems, not for resizing file
systems within a fixed boundary.  The difference is noticeable when
there is unallocated space because the file system doesn't fill the
whole disk device.  The dialog starts resizing a virtual partition the
size of the whole disk device, not the actual size of the file system.
Leave addressing this for a possible future update.

Bug 743181 - Add unpartitioned drive read-write support